examples/text-processor-service/requirements.txt
praisonai-svc>=1.2.0
python-dotenv>=1.0.0

